requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.15
alembic>=1.11.1
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
openai>=0.27.8
python-multipart>=0.0.6